Move the JS cache class stuff back into binding
